requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
python-jose[cryptography]
